Rename queue.py -> get_queue.py
