+++ tests/infrastructure/rq/test_queue.py
@@ -3,7 +3,7 @@ from redis import Redis
 from rq import Queue
 
 from erica.config import get_settings
-from erica.infrastructure.rq.queue import get_queue, QueueNotAvailableError
+from erica.infrastructure.rq.get_queue import get_queue, QueueNotAvailableError
 
 
 class TestQueue:
